It is compiling but not controlling right the quadrotors.
Finally I reached to parametrize the simulink model for N robots. The dimensions are working and it is compiling. Now the problem is that there should be some kind of error in the controller because the quadrotors are not controlled in the right way. The model with the for each block works but to me it looks way slower than the one with the parametrized controller. I don't know if this is just an impression or it is true. I will find out asap.

I renamed forEach folder files to avoid shadowing and other problems by MATLAB.