Parametrized controller is working.
Now I can simulate from 1 to N robots with the parametrized simulator. I don't know yet which is better, if the one with the for each or the one which is parametrized. I will try to figure it out but it is good that I have now both working.

The next step is to include a variable which contains the string from which the MATLAB understand what is the name of the robots. Probably from that string will be also possible to deduce n_robots, in this way the user has to change just one String everytime he/she wants to change the number of the robots. (Of course this tring has to be in sync with the name the user gives to his/her quadrotors objects in the V-REP scene).